+++ src/main/java/com/cafe24/nonchrono/controller/RecruitController.java
@@ -1,5 +1,6 @@
 package com.cafe24.nonchrono.controller;
 
+import com.cafe24.nonchrono.dao.RecruitDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
